feat: AI 베이스 모달 추가
