+++ src/common/component/AiFailModal/AiFailModal.tsx
@@ -1,7 +1,7 @@
 import * as styles from './AiFailModal.css';
 
+import AiModalBase from '@/common/component/AiModalBase/AiModalBase';
 import Button from '@/common/component/Button/Button';
-import { IcModalDelete } from '@/assets/svg';
 
 interface AiFailModalProps {
   onClose: () => void;
@@ -9,22 +9,11 @@ interface AiFailModalProps {
 }
 
 const AiFailModal = ({ onClose, message = '다시 한 번 시도해주세요.' }: AiFailModalProps) => (
-  <div className={styles.modalContainer} role="dialog" aria-modal="true">
-    <div className={styles.contentWrapper}>
-      <div className={styles.iconWrapper}>
-        <IcModalDelete className={styles.closeIcon} onClick={onClose} />
-      </div>
-      <div className={styles.textWrapper}>
-        <h2 id="ai-fail-title" className={styles.title}>
-          AI 추천 실패
-        </h2>
-        <p className={styles.description}>{message}</p>
-      </div>
-      <div className={styles.buttonWrapper}>
-        <Button text="다시 시도" onClick={onClose} />
-      </div>
+  <AiModalBase onClose={onClose} title="AI 추천 실패" description={message} titleId="ai-fail-title">
+    <div className={styles.buttonWrapper}>
+      <Button text="다시 시도" onClick={onClose} />
     </div>
-  </div>
+  </AiModalBase>
 );
 
 export default AiFailModal;
