+++ src/common/component/AiRecommendModal/AiRecommendModal.tsx
@@ -3,8 +3,9 @@ import { useState } from 'react';
 import * as styles from './AiRecommendModal.css';
 import Button from '../Button/Button';
 
+import AiModalBase from '@/common/component/AiModalBase/AiModalBase';
 import { usePostAiRecommendToCoreGoals } from '@/api/domain/upperTodo/hook';
-import { IcModalDelete, IcCheckboxDefault, IcCheckboxChecked } from '@/assets/svg';
+import { IcCheckboxDefault, IcCheckboxChecked } from '@/assets/svg';
 
 interface AiRecommendModalProps {
   onClose: () => void;
@@ -75,47 +76,45 @@ const AiRecommendModal = ({
   };
 
   return (
-    <div className={styles.container} role="dialog" aria-modal="true" aria-labelledby="modal-title">
-      <div className={styles.iconWrapper}>
-        <IcModalDelete className={styles.closeIcon} onClick={onClose} />
-      </div>
-      <div className={styles.contentWrapper}>
-        <h2 id="modal-title" className={styles.title}>
-          AI가 추천해 준 할 일이에요!
-        </h2>
-        <p className={styles.subtitle}>
+    <AiModalBase
+      onClose={onClose}
+      title="AI가 추천해 준 할 일이에요!"
+      description={
+        <>
           앞으로 <span className={styles.highlight}>{remainingSelections}개</span>를 더 선택할 수
           있어요
-        </p>
-        <div className={styles.listWrapper}>
-          {displayOptions.map((option) => {
-            const isChecked = selectedOptions.includes(option);
-            const isDisabled = !isChecked && selectedOptions.length >= emptyCount;
-            const CheckIcon = isChecked ? IcCheckboxChecked : IcCheckboxDefault;
+        </>
+      }
+      titleId="modal-title"
+    >
+      <div className={styles.listWrapper}>
+        {displayOptions.map((option) => {
+          const isChecked = selectedOptions.includes(option);
+          const isDisabled = !isChecked && selectedOptions.length >= emptyCount;
+          const CheckIcon = isChecked ? IcCheckboxChecked : IcCheckboxDefault;
 
-            return (
-              <div
-                key={option}
-                className={`${styles.listItem} ${isDisabled ? styles.listItemDisabled : ''}`}
-                role="button"
-                tabIndex={0}
-                onClick={() => {
-                  if (!isDisabled) {
-                    toggleOption(option);
-                  }
-                }}
-              >
-                <CheckIcon className={styles.checkboxIcon} />
-                <span>{option}</span>
-              </div>
-            );
-          })}
-        </div>
-        <div className={styles.buttonWrapper}>
-          <Button text="내 만다라트에 넣기" onClick={handleClick} />
-        </div>
+          return (
+            <div
+              key={option}
+              className={`${styles.listItem} ${isDisabled ? styles.listItemDisabled : ''}`}
+              role="button"
+              tabIndex={0}
+              onClick={() => {
+                if (!isDisabled) {
+                  toggleOption(option);
+                }
+              }}
+            >
+              <CheckIcon className={styles.checkboxIcon} />
+              <span>{option}</span>
+            </div>
+          );
+        })}
       </div>
-    </div>
+      <div className={styles.buttonWrapper}>
+        <Button text="내 만다라트에 넣기" onClick={handleClick} />
+      </div>
+    </AiModalBase>
   );
 };
